# docs/swagger.yaml
definitions:
...
  dto.IdentityChallengeWithEmailDTO:
    properties:
      email:
        type: string
      organization_id:
        type: string
    type: object
  dto.IdentityChallengeWithPhoneDTO:
    properties:
      organization_id:
        type: string
      phone:
        type: string
    type: object
...
  dto.IdentityRefreshTokenDTO:
    properties:
      refresh_token:
        type: string
    type: object
...
  dto.IdentityUserAuthDTO:
    properties:
      access_expires_at:
        type: string
      access_token:
        type: string
      last_login_at:
        type: string
      refresh_expires_at:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/dto.IdentityUserDTO'
    type: object
  dto.IdentityUserChallengeDTO:
    properties:
      challenge_at:
        type: string
      session_id:
        type: string
    type: object
  dto.IdentityUserDTO:
    properties:
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      full_name:
        type: string
      id:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      name:
        type: string
      phone:
        type: string
      status:
        type: boolean
      updated_at:
        type: string
      user_name:
        type: string
    type: object
  dto.IdentityUserLoginDTO:
    properties:
      organization_id:
        type: string
      password:
        type: string
      user_name:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      errors:
        items: {}
        type: array
      message:
        type: string
      status:
        type: integer
    type: object
  response.SuccessResponse:
    properties:
      code:
        type: string
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/v1/identity/challenge-verify:
    post:
      consumes:
      - application/json
      description: Verify the challenge
      parameters:
      - description: session_id
        in: path
        name: session_id
        required: true
        type: string
      - description: code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful verify the challenge
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Verify the challenge
      tags:
      - users
  /api/v1/identity/challenge-with-email:
    post:
      consumes:
      - application/json
      description: Login with email and otp
      parameters:
      - description: challenge payload
        in: body
        name: challenge
        required: true
        schema:
          $ref: '#/definitions/dto.IdentityChallengeWithEmailDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful make a challenge with Email and OTP
          schema:
            $ref: '#/definitions/dto.IdentityUserChallengeDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login with email and otp
      tags:
      - users
  /api/v1/identity/challenge-with-phone:
    post:
      consumes:
      - application/json
      description: Login with phone and otp
      parameters:
      - description: challenge payload
        in: body
        name: challenge
        required: true
        schema:
          $ref: '#/definitions/dto.IdentityChallengeWithPhoneDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful make a challenge with Phone and OTP
          schema:
            $ref: '#/definitions/dto.IdentityUserChallengeDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login with phone and otp
      tags:
      - users
  /api/v1/identity/login:
    post:
      consumes:
      - application/json
      description: Authenticate user
      parameters:
      - description: login payload
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.IdentityUserLoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful authenticate user
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Authenticate user
      tags:
      - users
  /api/v1/identity/login-with-apple:
    post:
      consumes:
      - application/json
      description: Authenticate user with Apple
      produces:
      - application/json
      responses:
        "200":
          description: Successful authenticate user with Apple
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Authenticate user with Apple
      tags:
      - users
  /api/v1/identity/login-with-facebook:
    post:
      consumes:
      - application/json
      description: Authenticate user with Facebook
      produces:
      - application/json
      responses:
        "200":
          description: Successful authenticate user with Facebook
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Authenticate user with Facebook
      tags:
      - users
  /api/v1/identity/login-with-google:
    post:
      consumes:
      - application/json
      description: Authenticate user with Google
      produces:
      - application/json
      responses:
        "200":
          description: Successful authenticate user with Google
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Authenticate user with Google
      tags:
      - users
  /api/v1/identity/logout:
    post:
      consumes:
      - application/json
      description: De-authenticate user
      produces:
      - application/json
      responses:
        "200":
          description: Successful de-authenticate user
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: De-authenticate user
      tags:
      - users
  /api/v1/identity/me:
    get:
      consumes:
      - application/json
      description: Get user profile
      produces:
      - application/json
      responses:
        "200":
          description: Successful get user profile
          schema:
            $ref: '#/definitions/dto.IdentityUserDTO'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get user profile
      tags:
      - users
  /api/v1/identity/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh token
      parameters:
      - description: refresh token payload
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/dto.IdentityRefreshTokenDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful refresh token
          schema:
            $ref: '#/definitions/dto.IdentityUserAuthDTO'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh token
      tags:
      - users
...
swagger: "2.0"